Move GetPeakLoad to executor package
GetPeakLoad which is defined in both Memcached and SPECjbb
sensitivity experiment shall be moved to pkg/executor package
