+++ experiments/specjbb-sensitivity-profile/common/common.go
@@ -3,7 +3,6 @@ package common
 import (
 	"github.com/intelsdi-x/swan/pkg/executor"
 	"github.com/intelsdi-x/swan/pkg/workloads/specjbb"
-	"github.com/pkg/errors"
 )
 
 // PrepareSpecjbbLoadGenerator creates new LoadGenerator based on specjbb.
@@ -40,32 +39,4 @@ func PrepareSpecjbbLoadGenerator(ip string) (executor.LoadGenerator, error) {
 		transactionInjectors, specjbbLoadGeneratorConfig)
 
 	return loadGeneratorLauncher, nil
-}
-
-// GetPeakLoad runs tuning in order to determine the peak load.
-func GetPeakLoad(hpLauncher executor.Launcher, loadGenerator executor.LoadGenerator, slo int) (peakLoad int, err error) {
-	prTask, err := hpLauncher.Launch()
-	if err != nil {
-		return 0, errors.Wrap(err, "cannot launch specjbb backend")
-	}
-	defer func() {
-		// If function terminated with error then we do not want to overwrite it with any errors in defer.
-		errStop := prTask.Stop()
-		if err == nil {
-			err = errStop
-		}
-		prTask.Clean()
-	}()
-
-	err = loadGenerator.Populate()
-	if err != nil {
-		return 0, errors.Wrap(err, "cannot populate memcached")
-	}
-
-	peakLoad, _, err = loadGenerator.Tune(slo)
-	if err != nil {
-		return 0, errors.Wrap(err, "tuning failed")
-	}
-
-	return
 }
